**Old impl:**
```
goos: linux
goarch: amd64
pkg: github.com/ak7sky/otus-go-prof/hw10_program_optimization
cpu: 12th Gen Intel(R) Core(TM) i9-12900H

BenchmarkGetDomainStatOld-20    	     489	   2237791 ns/op
BenchmarkGetDomainStatOld-20    	     486	   2278208 ns/op
BenchmarkGetDomainStatOld-20    	     484	   2388077 ns/op
BenchmarkGetDomainStatOld-20    	     444	   2343810 ns/op
BenchmarkGetDomainStatOld-20    	     475	   2317928 ns/op
BenchmarkGetDomainStatOld-20    	     436	   2446288 ns/op
BenchmarkGetDomainStatOld-20    	     456	   2317956 ns/op
BenchmarkGetDomainStatOld-20    	     452	   2311432 ns/op
BenchmarkGetDomainStatOld-20    	     476	   2297917 ns/op
BenchmarkGetDomainStatOld-20    	     470	   2283434 ns/op
PASS
ok  	github.com/ak7sky/otus-go-prof/hw10_program_optimization	53.816s
```
**New impl:**
```
goos: linux
goarch: amd64
pkg: github.com/ak7sky/otus-go-prof/hw10_program_optimization
cpu: 12th Gen Intel(R) Core(TM) i9-12900H

BenchmarkGetDomainStatNew-20    	 1170488	      1277 ns/op
BenchmarkGetDomainStatNew-20    	 1049586	      1097 ns/op
BenchmarkGetDomainStatNew-20    	  886833	      1398 ns/op
BenchmarkGetDomainStatNew-20    	  948170	      1717 ns/op
BenchmarkGetDomainStatNew-20    	  747936	      1628 ns/op
BenchmarkGetDomainStatNew-20    	  863659	      1719 ns/op
BenchmarkGetDomainStatNew-20    	 1006317	      1424 ns/op
BenchmarkGetDomainStatNew-20    	 1138748	      1369 ns/op
BenchmarkGetDomainStatNew-20    	  812144	      1242 ns/op
BenchmarkGetDomainStatNew-20    	  882260	      1558 ns/op
PASS
ok  	github.com/ak7sky/otus-go-prof/hw10_program_optimization	23.306s
```
**benchstat out:**
```
goos: linux
goarch: amd64
pkg: github.com/ak7sky/otus-go-prof/hw10_program_optimization
cpu: 12th Gen Intel(R) Core(TM) i9-12900H
                    │   old.txt   │       new.txt       │
                    │   sec/op    │    sec/op     vs base   │
GetDomainStatOld-20   2.315m ± 3%
GetDomainStatNew-20                 1.411µ ± 22%
geomean               2.315m        1.411µ        ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean
```